nuke cubit
