+++ lib/dartactyl.dart
@@ -7,4 +7,3 @@ export 'models.dart';
 export 'src/clients/application.dart';
 export 'src/clients/client.dart' hide PteroClientVisibleDio;
 export 'src/clients/translation_client.dart';
-export 'src/websocket/websocket.dart';
