+++ lib/websocket.dart
@@ -3,6 +3,4 @@ library dartactyl_websocket;
 
 import 'package:meta/meta.dart';
 
-export 'src/websocket_new/websocket_state.dart';
-export 'src/websocket_new/websocket_stream.dart';
-export 'src/websocket_new/websocket_stream_transformer.dart';
+export 'src/websocket/websocket.dart';
